remove add ap flag in right directory

+++ wifiAP/stopAP.sh
@@ -1,5 +1,11 @@
 #!/usr/bin/env bash
 
-rm /home/pi/data/wifiAP-enabled
+if [ -f /home/pi/data/wifiAP-enabled ]; then
+    rm /home/pi/data/wifiAP-enabled
+fi
+if [ -f /opt/openenergymonitor/data/wifiAP-enabled ]; then
+    rm /opt/openenergymonitor/data/wifiAP-enabled
+fi
+
 wifiAP stop
 sudo reboot
